requirements.txt
git+https://github.com/Rapptz/discord.py
ffmpeg-python
Wand